fix: organize commands into folders
