+++ src/main/java/frc/robot/autos/NewAuto.java
@@ -5,6 +5,8 @@ import frc.robot.subsystems.Swerve;
 import frc.robot.subsystems.Intake;
 import frc.robot.subsystems.Shooter;
 import frc.robot.commands.*;
+import frc.robot.commands.intake.RunIntakeWithTime;
+import frc.robot.commands.swerve.MoveToPose;
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
