+++ src/main/java/frc/robot/commands/intake/IntakeControl.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.intake;
 
 import java.util.function.DoubleSupplier;
