+++ src/main/java/frc/robot/commands/intake/IntakeTest.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.intake;
 
 import frc.robot.subsystems.*;
