+++ src/main/java/frc/robot/commands/shooter/PrimeAndShoot.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.shooter;
 
 import java.util.function.DoubleSupplier;
