+++ src/main/java/frc/robot/commands/swerve/MoveToPose.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.swerve;
 
 import java.util.List;
 import java.util.Optional;
