+++ src/main/java/frc/robot/commands/swerve/TeleopSwerve.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.swerve;
 
 import frc.robot.Constants;
 import frc.robot.subsystems.Swerve;
