+++ src/main/java/frc/robot/commands/swerve/pid/DriveToPose.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.swerve.pid;
 
 
 import edu.wpi.first.math.controller.PIDController;
@@ -13,17 +13,18 @@ import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.Constants;
 import frc.robot.subsystems.Swerve;
 
-public class Drive extends Command {
+public class DriveToPose extends Command {
     private Swerve s_Swerve;
     private Pose2d targetPose;
 
     private PIDController xController;
     private PIDController yController;
     private ProfiledPIDController driveController;
+    private ProfiledPIDController thetaController;
 
     private boolean isRunning;
 
-    public Drive(Swerve s_Swerve, Pose2d targetPose) {
+    public DriveToPose(Swerve s_Swerve, Pose2d targetPose) {
         this.s_Swerve = s_Swerve;
         this.targetPose = targetPose;
 
@@ -38,12 +39,18 @@ public class Drive extends Command {
 
         this.xController = new PIDController(Constants.AutoConstants.kPXController, 0, 0);
         this.yController = new PIDController(Constants.AutoConstants.kPYController, 0, 0);
-        this.driveController = new ProfiledPIDController(Constants.AutoConstants.kPXController, 0, 0, new TrapezoidProfile.Constraints(
-            Constants.AutoConstants.kMaxSpeedMetersPerSecond, Constants.AutoConstants.kMaxAccelerationMetersPerSecondSquared
-        ));
+
+        // this.driveController = new ProfiledPIDController(Constants.AutoConstants.kPXController, 0, 0, new TrapezoidProfile.Constraints(
+        //     Constants.AutoConstants.kMaxSpeedMetersPerSecond, Constants.AutoConstants.kMaxAccelerationMetersPerSecondSquared
+        // ));
+
+        this.thetaController = new ProfiledPIDController(
+            Constants.AutoConstants.kPThetaController, 0, 0, Constants.AutoConstants.kThetaControllerConstraints);
+        thetaController.enableContinuousInput(-Math.PI, Math.PI);
 
         Pose2d currentPose = s_Swerve.getPose();
-        s_Swerve.resetOdometry(new Pose2d());
+
+        s_Swerve.resetOdometry(currentPose);
 
         // driveController.reset(
         //     new TrapezoidProfile.State(currentPose.getTranslation().getDistance(targetPose.getTranslation()),
@@ -56,12 +63,13 @@ public class Drive extends Command {
         //             ).getX()
         //     )
         // );
+
+        thetaController.reset(currentPose.getRotation().getRadians());
     }
 
     @Override
     public void execute() {
         Pose2d currentPose = s_Swerve.getPose();
-        System.out.println("Drive: " + currentPose);
         
         double currentDistance = currentPose.getTranslation().getDistance(targetPose.getTranslation());
         double driveErrorAbs = currentDistance;
@@ -74,20 +82,10 @@ public class Drive extends Command {
                 currentPose.getTranslation().minus(targetPose.getTranslation()).getAngle()
             ).transformBy(new Transform2d(driveVelocityScalar, 0.0, new Rotation2d())).getTranslation();
 
-        System.out.println("Combined controller: " + driveVelocityScalar + ", " + driveVelocity);
+        // double thetaVelocity = thetaController.atGoal() ? 0.0 : 
+        //     thetaController.calculate(currentPose.getRotation().getRadians(), targetPose.getRotation().getRadians());
 
-        double xVelocity = xController.calculate(currentPose.getX(), targetPose.getX());
-        double yVelocity = yController.calculate(currentPose.getY(), targetPose.getY());
-
-        System.out.println("Before: " + xVelocity + ", " + yVelocity);
-
-        if (xController.atSetpoint()) { xVelocity = 0; }
-        if (yController.atSetpoint()) { yVelocity = 0; }
-
-        System.out.println("After: " + xVelocity + ", " + yVelocity);
-
-        // s_Swerve.drive(driveVelocity, 0, true, true);
-        s_Swerve.drive(new Translation2d(xVelocity, yVelocity), 0, true, true);
+        s_Swerve.drive(driveVelocity, 0, true, true);
     }
     
     @Override
@@ -97,7 +95,7 @@ public class Drive extends Command {
     }
 
     public boolean atGoal() {
-        return isRunning && driveController.atGoal();
+        return isRunning && driveController.atGoal() && thetaController.atGoal();
     }
 
     public boolean isRunning() {
