+++ src/main/java/frc/robot/commands/swerve/pid/Spin.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.swerve.pid;
 
 
 import edu.wpi.first.math.controller.PIDController;
